// src/app/model/TileNameEnum.ts
export enum TileNameEnum {
    t4 = "T4",
    i = "I",
    i2 = "I2",
    i3 = "I3",
    i4 = "I4",
    i5 = "I5",
    t5 = "T5",
    l4 = "L4",
    l5 = "L5",
    v3 = "V3", 
    v5 = "V5",
    z4 = "Z4", 
    z5 = "Z5",
    y = "Y",
    o = "O",
    u = "U",
    x = "X", 
    f = "F", 
    w = "W",
    n = "N", 
    p = "P"
}